CreateDmg: add jdkDependency
Since the JDK in mbeddr projects is downloaded from Nexus, it's simpler
to allow the user of CreateDmg to point to the JDK using the dependency
notation. This avoids the need to create and resolve a special
configuration just for the JDK.

As both dependencies and files may be specified as strings, a separate
setter is needed to distinguish between the two cases.