seperate project for loading idea applications
Moved all the logic required to start up a idea environment into the
`project-loader` subproject. The subproject contains all code required
to load plugins and open a project. The loader also takes care of the
complete lifecycle of the idea application and the project. Users of The
loader pass a lambda/function that takes the project as an argument. In
case of any errors during the execution of the lambda or when the lambda
is executed successfully the loader tries to gracefully shut down the
idea environment.

In addition to that the loader specifies a basic command line interface
for potential users of the loader.

The `execute-generators` subproject is the first user of the new
`project-loader`. Since it required changes to the command line
interface this is a breaking change.