add parameter to override the version
When using the tests we need to know the version ot the plugin we are
testing against statically, because there is no code can run before
the plugin block of a build script is evaluated. I added parameter to
force a snapshot version that we can then use in the tests.