update to MPS 2018.3
Updated the dependency on MPS to MPS 2018.3.3 from MPS 2018.2.6.
This updadate required us to use kotlin 1.3 because the intellij
platform migrated to that kotlin version. Kont 1.3 brings a stricter
null checker so we had to adapt some code in the helpers that interact
with the intellij platform.

Removed the deprecated `getInstance` call for the `FacetRegistry` with a
proper `Project.getComponent` call.