modelcheck: add support for JUnit XML report
The model check plugin can not report its results as a JUnit xml file
which seems the defacto standard for test reporting in CI. For each
model that is checked it will create a testcase entry and report if as
failed if any model checking error is reported in that model. If multiple
error are reported in the same model they will all get reported as a
failed assertion in that test case.

Reporting is controlled via the new `junitFile` configuration for the
plugin. If the configuration is present the grade plugin will report
into that file.