generate project libs: make overrides file optional
Make overrides file optional again. Seems in recent gradle versions the
behavior has changed and the overrides file throws an exception if it is
none existing. The file was anyway registed as  an input when it existed
in the setter so the InputFile annotation was not required.

Added tests for the tasts for both cases where the overrides file is
present and when it is not.