RunAntScript: inherit gradle log level
adds support for the new "mps.ant.log" property to the RunAntScript tasks.
The property is used in MPS build scripts generated with MPS 2020.2+
to set the log level during generation. For older versions it is
simply ignored.

The logic in RunAntScript checks if the current log level for the task
is set to anything different than "LIFECYCLE" which is the default and
sets property accordingly when it detects a none default level.

This allows settings the log level via gradle either via command line
e.g. -w, -i or -d or on task level in the build script.