generate, modelcheck: allow using custom MPS
Adds support for the custom distributions of MPS vai new property
mpsVersion which is used to resolve the runtime dependency.

Cleanup the code responsible for resolving MPS and removed the logic
to handle an edge case in 2019.2. This causes the plugin to be no longer
compatible with MPS versions lower than 2019.3. The plugin checks the MPS
version and presents an error is its older than 2020.1. The minor version
of the plugin was increased because of that.

Tests were added that contain test project for modelcheck and generate
tasks. One test is ignored because resoving the runtime dependency is still
tricky and not 100% solved. Never the less I wanted to get the feature into
review this week. We can merge it without that test case for now and
revamp the versioning of the runtime part later on.

The readme contains an overview how users can use their custom MPS.