Add `incremental` flag to RunAntScript, bump minor version
Having an explicit flag for controlling incrementality makes it a little
easier to use. It also lets us remove the `clean` target from the
invoked targets, this is something that those who configure incremental
builds manually often forget, thus making their builds not quite as
incremental as could be possible.

Since this is a new feature, the minor version is increased.